+++ changelog.md
@@ -7,6 +7,7 @@
 - Updated Node.js support to `^14.17.0 || ^16.0.0 || >= 18.0.0`.
 - Updated dev dependencies, some of which require newer Node.js versions than previously supported.
 - Use the `node:` URL scheme for Node.js builtin module imports.
+- Migrated from the Node.js builtin module `fs` to `node:fs/promises`.
 
 ### Patch